+++ src/main/java/com/ssafy/sandbox/todo/dto/TodoListResponseDto.java
@@ -1,6 +1,6 @@
-package com.ssafy.sandbox.dto;
+package com.ssafy.sandbox.todo.dto;
 
-import com.ssafy.sandbox.entity.Todo;
+import com.ssafy.sandbox.todo.entity.Todo;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
