+++ src/main/java/com/ssafy/sandbox/todo/repository/TodoRepository.java
@@ -1,7 +1,7 @@
-package com.ssafy.sandbox.repository;
+package com.ssafy.sandbox.todo.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.ssafy.sandbox.entity.Todo;
+import com.ssafy.sandbox.todo.entity.Todo;
 import org.springframework.stereotype.Repository;
 
 @Repository
